requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==25.3.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
discord.py==2.5.2
exceptiongroup==1.3.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.6.3
mysql-connector-python==9.3.0
ollama==0.6.0
prawcore==2.4.0
propcache==0.3.2
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
PyNaCl==1.5.0
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.14.1
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.8.0
yarl==1.20.1
